Use the ynabToolKit script and mutationObserver
Because the firefox build loads js files out of order, I _still_ need to
setup DOM polling on each script until ynabToolKit is defined.
Thankfully, once the page loads once, we are done with the polling.

So, with that work done, the following features can be enabled and
disabled independently and all use a single mutation Observer

- Check credit balances
- Highlight negatives negative
- Remove zero/negative categories

This works in Chrome and Firefox, untested in Safari.