Add feature for pacing
Pacing is calculated by determining how far have progressed through the
month and evaluating how much of the budget has been consumed. As an
example, if we are on the 15th of the month and have spent $60 of a $100
budget, we are $10 overpace. This allows us to course-correct midway
through the month.

Needs adjustments to work with other months and not repeatedly repaint
the DOM.