Stop-gap refactor of actOnChange
When a change happens in YNAB, the mutation observer might find multiple
changes to nodes that have the same class name. We only need to act on
the first instance of a relevant mutated node during each mutation event
and we should ignore the duplicates. 

However, previously, we would fire our feature scripts for each node in
each mutation event, regardless of duplicates.

This can and will be improved further, but now we loop through the
entire digest of changed nodes and only fire our feature scripts once
for each mutation event, vastly improving performance for features like
Pacing.