Refactored shared functions into a specific shared object on ynabToolKit
Because the ynabToolKit object is now created with the injectJSScript
function, the new function() object in shared/main.js needed to be
refactored.

Since shared functions are now available one step in (e.g.
ynabToolKit.shared.functionName), we needed to update the references of
those shared functions in the files that currently use them.