Now using async process to load and save options with promises.
Kango docs say that the API is only available async from the options page. Although it works on Chrome and Firefox with direct calls, it doesn't on Safari, so the refactor is necessary to ensure it works across all supported browsers. I was using a setTimeout hack which worked sometimes on Safari, but not all the time. This seems to work more consistently.