Balance to Zero fixes
Updates to make this feature properly pop up a warning if the Warn on
Quick Budget option is enabled. Hid the button when the category is
already at zero. Added a ‘+’ on positive amounts, to help differentiate
the functionality from other quick budget options (core YNAB quick
budget options replace the budgeted number with a new number, whereas
this adds or subtracts to what is already budgeted).