Fix Paid in Full Credit Card Assist
Fixes Paid in Full Credit Card Assist when using non-default number formats.
Eg. when using number format "123 456,78" the current implementation tries to set the field value to 1234.56 which then gets changed to 123 456,00.
Tested this with a couple different formats and it works with all of them, even with weird formats with slashes as decimal separators.