Must set record height if activity-transaction-link feature is active.
If the activity-transaction-link feature is not active we don't want to adjust the record height because doing so causes scrolling to be jumpier than usual. If the activity-transaction-link feature is active we do need to set the record height because it uses the value to scroll the "selected transaction" to the top of the register.