Changes related to switching budgets and uncategorized transactions.
Budget progress bars needed to be aware of switching budgets and reloading the categories. Added onRouteChanged() to accomplish that.

Budget category info (which also feeds the goal indicator feature) had a failing compare when the sub-category was Uncategorized Transactions. Added a .not() to the selector that gathers all the categories to ensure the Uncategorized Transactions category isn't processed.