Handle different currency formats correctly when doing math.
Changed the code to unformat() values before performing any math on it. One thing to note; switching to a budget that uses a different currency format (not real likely unless you're a developer), the decimal separator, and the number of fractional digits will not be set properly. THis will be addressed in a future code change.