Added a way to migrate legacy keys source keys to new sauce keys
This feature works by modifying `ensureDefaultsAreSet` to check
a map of keys which are stored in `allSettings`. If there is a mapped
version of a key, we check for the setting on that key. If it exists
we use it as the default otherwise we use the feature default.

This code should only run the first time a user opens the options page
or YNAB itself. Afterwards the key should exist and have the proper value.