Handle new function for transaction import
New bank data provider seems to have introduced a new function for determining if import is active for an account. Updated code to check for old function name and new function name and proceed if either exist and returns true. 

Also added timestamp to the build file to make it easier to know when builds start/finish. Mostly needed due to the double build requirement which will eventually be removed. :)