Goal Progress Bar Fixes
Fixed an error related to YNAB’s credit card functionality that stopped
this functionality from working. When selecting a category, the default
selection background is restored, rather than the previous highlighting
of just the category name, which didn’t always consistently work. Fixed
a bug with the % bar showing an incorrect percentage for monthly goals.
Called this feature a little more often, so the bar will update after a
budgeted value is changed.