Budget Category Info Fix
Corrected a bug where the “Highlight all negatives red” feature and the
Pacing feature conflicted, potentially resulting the labels in the
Available and Pacing columns changing colors when a category was
selected. Also ensured that the Budget Category Info loop ran when
switching from another section of the app to the Budget section, so
features that rely on that code will work more consistently.