Fix retro calculator feature
The feature was only worrying about the `budget.select` route which
is specific to budget pages with a month specified. But it's possible to
get to the `budget.index` route by navigation from open budget -> your budget.
This result then gets cached apparently in local storage and used which is
weird but...there ya have it.

Now we just check for the word `budget` in the route name which means it
will run on the "Open A Budget" page but that should be okay since the
keypress listener is added to the document which will be there on that page :D