Potential fix for running balance inconsistencies
I've gone with the assumption that somehow accoutns get reset (switching
budgets, maybe?) which causes our observer on `anyItemChangedCounter` to go
away. This change will try and keep track of that happening. Unfortunately,
if that observer gets wiped out but the accoutn object itself doesn't get
overwritten this fix won't actually work. Let's hope that's not the case (: