Refactor ensureDefaultsAreSet to getUserSettings
Instead of just making sure defaults are set, getUserSettings now ensures
defaults are set and returns a settings object with all the user-specified
settings as well. The logic to migrate extension setting names and to ensure
users settings are stored in browser.storage is now also part of calling
getUserSettings.