Upgrade ESLint and HighCharts
Upgrading eslint to get rid of some build warning since we're so out
of date. Ran fix for all the reasonable new rules that were added to
the airbnb preset and turned off the ones that would have been a major
hassle without unit tests. Also updated highcharts, fixed a modal issue
on the SpendingByCategory report, and added docs for how we vendore the
HighCharts library.