Wrap lifecycle calls with a try-catch
Created a withToolkitError wrapper function which you can use to wrap
a call to some function with a try-catch which will verbosely log an error
to the console. In the future we can use this wrapper to also increment a
counter on top of the icon in the browser informing the user an error has
occured and perhaps provide a way for them to automatically report the error
to us.