Wait for budgetVersionId to change before calling onBudgetChanged
We attach an observer to the applicationController's budgetVersionId
to detect when the user has switched budgets but this value changes
before YNAB loads data causing our feature code to potentially operate
on stale data. Now, we stil listen to the field because it's the best
source of truth we have but once it changes, we begin polling for the
budgetViewModel's budgetVersionId to match -- this should always happen
because the budget is the first view users see once swapping budgets.