Initialize the running balance for all accounts in the new invoke method
Before, we were only initializing running balance in willInvoke this
caused a bug where the code would run before you've selected a budget
(on the users/budgets page). Now, we will always attempt to initialize
the running balance on accounts whenever the feature is invoked which
happens on route changes. We've included a guard preventing this function
from processing once it's already initialized.