Invert stealing from future month date logic
We were only considering dates in the future becasue that's what the feature
is for but it's still of course possible to overbudget in the current month
while having this feature enabled. We'd show you that you're stealing from the
next month but in reality, you're just overbudget in the current month. Inverting
the logic to check we're not looking at a past month fixes this issue.