Refactor pif assist feature to generically add styling
We were applying styles directly to elements that cared about them
which made it difficult to play well with other features. Now we
add a data-* tag to the root category column and inspector elements and
use CSS to style when those data tags are there. Also refactored out
use of the budgetView which should resolve some Sentry errors :D