Run content scripts at `document_idle`
I'm hoping this fixes the issue where YNAB loads with a missing
sidebar. If this doesn't solve it, we may want to look into using
requestAnimationFrame before injecting our stuff into the DOM. This
issue is definitely some sort of race condition so I'll keep playing
with timing related fixes until we hopefully find a fix.