Update styles for remove-positive-highlight to not hide icons
The hope was to rewrite this using SASS but unfortunately, SASS injects
styles into the bundle which means we can't be smart about when we load
it, it's always loaded. There may be things we can do in the future
to make it so features can use SASS but until then, I'll just commit the
compiled css.