Fixed an issue where "Unhighlight Positive Inflows" hid the new budget page icons added by YNAB. (#1448)
* Remove legacy toolkit report classes, add currency to react-reports

* Adding some reusable color variables

* Update styles for remove-positive-highlight to not hide icons

The hope was to rewrite this using SASS but unfortunately, SASS injects
styles into the bundle which means we can't be smart about when we load
it, it's always loaded. There may be things we can do in the future
to make it so features can use SASS but until then, I'll just commit the
compiled css.