:see_no_evil: hide yo project time
This is nothing special.  These are just two common meta files.
The purpose of them is for identifying a project to a time tracking srvc

I appologize for the late submission on this.
but if i'm going to work on a project,
then having these files ignored would be nice