:sparkles: splitting chrome and firefox features
for now we will split these into two seperate features.
I will continue to search for a better solution, howerver,
releasing this pr while we search for it gets this functionality
out to the people at least.
It should close #2062
If you accept this PR, can you suggest a tag to use
for discussing this browser issue in the issues section?