issue-2177: Fix report bug button in the extension popup not opening in Firefox
From what I can gather, this happens because the `window.close()` causes the script to unload and the browser isn't quick enough to open the tab. Adding a 50ms delay was enough in my tests for the tab to open and it is still short enough that it would not be a noticable delay.