fix(budget): Change vis. of card in privacy mode
- Change visibilty of currency numbers in the Live on last month's
income while in privacy mode.

The values in the "Live on last month's income" Card on the budget
screen did not have the global currency class attached.
Therefore, when privacy mode is enabled, the changes made for privacy
mode could not be applied to the values in the card.

This fix is to add the global currency CSS class to the values in the
card to allow privacy mode to properly work.

Closes #2686.