fix!: update top-accounts lookups (#2889)
YNAB pushed updates on 7/25/22 that broke the `top-accounts`
Ember component. Internally, we look for this component only to get
the `areChecked` property, to find checked transactions
(and once to setup an Observer on that property).
This commit changes all `componentLookup('top-accounts')`
to `containerLookup('service:accounts'), as it shares that property.

BREAKING CHANGE: This breaks the SpareChange feature. It looks like the
Observer fails to be added correctly. I think this is still worth pushing for now,
though, as the feature would still be broken without this and this repairs
the other uses of `top-accounts`