fix(normalizeNetIncomes): Handle edge cases
There are some instances where expenses and incomes monthly totals don't
start or end on the same months. The case where they might start on
different months can happen if you create an account, then add some
transactions to it with dates in prior months to the month the starting
balance was set it. The case where they might end on different months may
not actually exist, since latest months should always have at least an
empty array of totals for both. But it can't hurt to handle it just incase.